fixed first msg to inactive bkgd failing ... and it was a long journey
The service worker is absolutely required to register and activate its
listeners *immediately* at boot time, or it will fail to handle the
event(s) which woke it up.  This means those must be active before it
tries to do any async operations -- like loading config from storage.

So then it needed a mechanism to delay handling of events until the
init had finished, like if it's awakened by a request for a new Node ID,
it can't generate an ID until it knows its own client ID, which it can't
know until waiting for async I/O from the storage API to load config data.
This means the events must be handled asynchronously.

BUT

The runtime.onMessage event cannot be handled async.  If you receive a
message and then do some async stuff, it'll close the message channel
before the async stuff is done, and then it can't send a response.

So it must queue or spawn an async handler and then return `true`
immediately, which tells the message API to keep the channel open for
a reply later.  This also means a reply is now mandatory, and it'll
cause errors if the reply doesn't happen within a few seconds.  And when
an error happens, my emit() call retries... which can have some really
bad side effects in some cases.  So it needs to make sure to avoid that.

I had to rewrite the Bkgd.onMessage handler to work in this different way,
and it took a lot of experimenting to find a solution which didn't have
problems.  At least, no problems that I was able to find, anyway.

As for waiting for config to load, and other slow tasks during init,
I had to add Promises which resolve only when that part of the init
is complete.  Then make event handlers wait on those promises before
executing their own code.

The code changes, in a bit more detail...

Bkgd:
- registered Bkgd listeners immediately during init, to avoid missing
  the first instance of the event which woke up the script
- made init load config async, then load the tree async, and resolve
  a promise for each when it's complete
  (added Bkgd.configLoaded and Bkgd.treeLoaded Promises)
- made Bkgd.onMessage return true immediately if the message is for bkgd,
  so it can have time to process slow operations before responding
- made Bkgd.onMessage ignore all other events, because any response
  to those causes problems
- changed Bkgd.bkgd_*() API to return a value to the dispatcher
  instead of calling sendResponse() in the individual handlers,
  and the dispatcher ALWAYS sends a response, no matter what
  (this avoids problems and simplifies the code)

Common:
- made emit() validate its arguments better, to avoid hard-to-debug issues
  when the args were incorrect
- reduced emit() max tries since previous value was so big it was
  impossible to reach before timeouts interrupt it
- replaced most log() events with debug(), to reduce log clutter

Tree:
- Tree.onMessage() now correctly ignores messages not intended for it

NodeView:
- NodeView.newNodeID() uses emit() now (I forgot to convert it earlier)