added node timestamps and node details box...
For timestamps, I had to change how Node events are passed, to include
the `msg` dict in all calls next to `notify`, and add the timestamp
as `msg.when`.  Had to update a lot of function signatures for this.

The reason is to allow the timestamp to be part of the same node change
event, instead of "bump timestamp" being a separate event.  I want them
bundled together to reduce the transaction log entries and make undo
easier later.  Plus, it just makes sense to have a timestamp on all
events of this nature, even in cases when the timestamp might not be used.

Node timestamps are mostly similar to posix filesystem timestamps, but
not identical.  The ctime is creation time, not change time.  The mtime
doesn't track attributes which won't be synced to the server, like
"marked".  The atime is used for expanding/collapsing a node, even
though that technically could be mtime.  I'm still working out what the
semantics ideally should be, and whether it needs any addition stamps
for things like the last time a tab was loaded.

The node detail box displays stuff which can't reliably be fully seen
in the tree view.  It can be turned on and off with the new "Details"
button, but the state isn't persisted yet.

misc other stuff
- added common:fmtDate() to convert millis to ISO8601-style date
  (but in the user's local timezone, without the "T" marker)
- simplified Node.toDict() and Node.fromDict()
- removed "Window ${window.id}" generated note from new windows,
  since it's redundant
- added a "Window" or "Session" scope button for later use
  (whole session mode, or show only this window)
- moved bottom area top border to #details-box or #bottom-bar
  depending on whether #details-box is visible
  (I tried putting them both in a shared container div, but that had
  unwanted side effects since #details-box can scroll horizontally
  and the other bottom elements shouldn't)
- fixed text selection cursor being displayed over clickable buttons
- added a style for buttons which stay pressed