partially implemented onTabRemoved and tab unloading / closing and loading - actually closes tabs when a loaded Node gets deleted - added logic to unload or delete based on Node and Tree state   (like, unload if it has notes, delete if there's no metadata to save) - started adding Node.load() to restore unloaded tabs,   but it's barely there and still needs a LOT of work - made Node.unload() actually work, pretty much   (but it doesn't handle window Nodes yet) - fixed bug: Node.unload() didn't set active=false   (so unloaded active tabs would remain "active") - added 'u' key to unload cursor node - added 'Enter' key to load or edit cursor node
Some double-event errors are starting to pop up, and I'm not sure
how best to handle or prevent them.  For now, it's checking for known
expected "error" conditions and ignoring them... but it feels dirty.
For example:
  - put cursor on a loaded tab in the TreeView
  - press 'd' to delete the node
  - node gets deleted, and closes the tab
  - browser triggers the onTabRemoved event
  - onTabRemoved tries to delete the node again

Another example:
  - close the active tab with C-w in the tab
  - closed tab Node tries to setActive(false) after it has been deleted

It seems like it might help to have some sort of global state
synced between all threads, with stuff like "that deleted tab was
because of me, don't worry if you can't find it".  But that seems like
it'd just run into *different* problems... so I feel like maybe I should
be using an entirely different approach, but I'm not sure what.