fixed bug loading saved tab: infinite loop
Or at least, I think I fixed it.  Not sure what caused it, exactly,
but I haven't been able to make it happen again after this change.

It'd get infinite onTabUpdated events,
going between 'loading' and 'complete',
but only sometimes, and I'm not sure what conditions caused it.
Regardless, this seems to end the loop
and saved tabs can still be loaded.