made load/unload window work - implemented Node.unload() for windows - implemented Node.load() for saved windows - made 'onWindowClosed' become 'onWindowRemoved', for consistency   with the browser's event names
It even works if the window nodes are nested.  Loading/unloading a
window only affects that window, not any nested windows within.
The requested window comes right back in the same size and position
it was in before closing, and opens all the same tabs it had before.
The 'wasLoaded' state determines which tabs are opened, so it doesn't
go crazy and try to load ALL the tabs... just the ones which were open
at the time the window was saved.