fixed firefox manifest so it can actually be loaded (before, it only worked as a "temporary extension")
Also, added a placeholder for a feature I want to add later:
A hotkey the user can press to add any currently-highlighted text
to the current tab's note.  So while reading a page, they could
highlight any important bits and press a key, and append/prepend it
to that tab's note field.