made checkboxes recalculate when nodes move, and cleaned up API a bit - added Node.hasCheckbox() to isolate the logic for that check - added Node.getCheckboxType() to isolate that logic too - added Node.hasKidsWithCheckboxes() - made previous and new parent recalculate during Node.moveTo - made Node.updateCheckboxes() start with self, not parent   (to make previous parent actually able to recalculate itself) - made 'percent' types always return their percent,   so other types without kids can avoid returning stale percents
The upside of this change is that the checkbox completion amounts
update automatically while moving nodes around.

The downside of this change is that the checkbox completion amounts
update automatically while moving nodes around.  Specifically, if you
hold Shift+Up/Down to move a done / half-done / todo node around a
bunch of times, it can destructively change the completion state of
the nodes it passes by along the way.  By becoming a child of those
nodes, it changes their completion and it doesn't always change back
when it leaves.

This seems to be fairly unavoidable, so I'm still debating whether
the completion info should be updated during move or not.  If not,
this problem goes away... but then the user must manually trigger
an update to recalculate the percents after a move.  So I think I'll
maybe add a user config option and let the user decide.