disabled "Add Note" css on incoming text drag, because browsers hate it
Firefox works, but everything else tends to spaz out when that
pseudo-element is added during a drag, and they lose their drag target,
which makes the element disappear, which restores the drag target,
which repeats the cycle.  It's nearly impossible to drag to there.

But without that css, it works fine.