fixed a few things... - fixed tab reorder storms again - reduced known errors from moving tabs - fixed some cases when multiple tab nodes could end up marked as "active" - fixed loading saved windows with forbidden geometry
I still can't get Brave to work in every case, because it does some
*wild* shit... but this seems to work reliably in Firefox and Vivaldi.
Brave has multiple unfortunate behaviors when dragging a tab in the tab
bar, or dragging a tab *out* of the tab bar.  Like, on drag-out, it'll
create a new window immediately, attach the tab to it, then delete the
new window a moment later, attach the tab to its original window but in
position 0 (for no apparent reason), then create a *second* new window,
attach the tab to it, and then finally settle.  Meanwhile, it creates
an absolute storm of events, *and* it rejects responses to those events
during the drag.

So I had to add new mechanisms to handle that.  It keeps retrying when
that error happens, until it succeeds or until the timeout expires.
It assumes the user won't hold a single drag for more than 30 seconds.

Bkgd:
- fixed bug: loading saved window failed if user saved a bad position
  (Chrome has rules about where it'll allow new windows to be opened,
  and the user can save+close a window in a forbidden position...
  which would cause an error when trying to restore the window)
- removed the need to lock before onTabMoved()
  (we actually want events to overlap now, for debounce purposes)
- rewrote bkgd_reorderAllTabsInThisWindow().  No need for a mutex any
  more.  Instead, it debounces incoming requests and executes a delayed
  call to the actual Node.reorderAllTabsInThisWindow function.  This
  seems to work in more cases, and it prevents infinite loops.

Tree:
- made Tree.onTabActivated() async
- made Tree.onTabAttached() move nodes directly
  instead of calling onTabMoved(),
  because the delegation was making things weird and fragile
- made onTabAttached() also ensure only one tab is 'active'

Node:
- fixed silent failure in Node.moveTo() when no args given
- added 'onTabAttached' as a "reason" for some events
- made Node.setActiveTab() async and returns whether any changes were done
  and avoids doing anything when no change was needed
- Node.reorderAllTabsInThisWindow() uses a totally new algorithm,
  which handles the weird stuff Brave does during a tab drag
- handle/ignore "Tabs cannot be edited right now" errors in Brave