made Node.setActiveTab() MUCH more reliable by debouncing and delaying until new requests stop coming in
It doesn't make any changes until it has been 100 ms since the most recent
request.  Then it queries the browser to find the active tab, and updates
the tree accordingly.  This prevents a bunch of issues, and handles
things better even when the tree is in a weird state, by correcting the
state of the tree.

The 100 ms value was chosen because it's less than most users' key-repeat
rate, and short enough to not feel super laggy, but long enough to avoid
any of the prior issues I was encountering.

So if the user holds Ctrl+PgDn to scroll rapidly through tabs, it waits
until the user lets go before updating the tree.

It also responds correctly if the user presses the key exactly 100ms apart
(which I tested quite a bit in each of my test browsers).