reversed the numeric task colors, to emphasize 1 instead of 9
Could go either way on this, really.  Priority 1 is important, if the
number represents a priority... but if it represents a score or weight,
then 9 would be important instead.  It really just depends on what
meaning the user wants to assign to it.

But for tasks, I think "1 = important" makes more sense as a default.
Because then the numbers could indicate priorities, or the order of steps,
or the completion level.

The original coloring seems better suited to archival than to tasks.
Like, rating a movie on a scale of 1 to 9, 9 is better and brighter.
But I don't think this type of use will be as common.