fixed warning during onTabRemoved after node.unload()
It would look up the tab and be unable to find it, because the tabId
had already been unset.  So this holds onto that value very briefly
and calls unload() a second time when the tab has actually been closed,
to finalize the unload.

Node.unload(userAction) tells the browser to close the tab,
then onTabRemoved() calls Node.unload(onTabRemoved)
to detach the node from the tab.
... while tree_nodeChanged(unload) keeps threads synchronized.

Also fixed an earlier bug which wasn't possible to trigger until now,
which would allow non-userAction events to toggle the wasLoaded state.