during fsck, re-attach orphaned nodes under "lost+found" instead of deleting
This is similar to how some filesystems attach orphaned inodes under a
"lost+found/" directory during fsck, to give the user a chance to recover
any data which got unlinked in unintended ways.

It's probably better to save this stuff to be safe, instead of deleting
any nodes without parents.  Just in case something got corrupted or
whatever, this should greatly reduce the amount of data lost.  The user
can delete it later if they want.